update group creation repository
