+++ app/src/repository/GroupCreationRepository.js
@@ -6,15 +6,25 @@ import GroupingCreationDto from '../dto/GroupingCreationDto';
 
 const TARGET_URL = `${SERVER_URL}/group`;
 
-const GroupCreationRepository = () => {
-  const completeGroupCreation = async (groupingCreationDto) => {
+export default class GroupCreationRepository {
+  async completeGroupCreation(groupingCreationDto) {
     try {
+      console.log('group creation complete');
       return await axios.post(`${TARGET_URL}`, groupingCreationDto);
     } catch (error) {
       console.error(error);
     }
     return null;
-  };
-};
+  }
 
-export default GroupCreationRepository;
+  async completeGroupRepresentImg(groupId, imageFile, failedCallback) {
+    const response = await axios.post(`${TARGET_URL}`, groupId, imageFile);
+    console.log('response');
+    console.log(response);
+    const commonResponse = new CommonResponse(response.data);
+    if (commonResponse.code !== ResponseCode.SUCCEED) {
+      failedCallback(commonResponse.code);
+    }
+    return new GroupingCreationDto(commonResponse.data);
+  }
+}
